requirements.txt
aiohttp==3.8.5
pyyaml==6.0.1
